requirements.txt
blinker==1.9.0
click==8.1.8
Flask==2.3.3
Flask-Cors==4.0.0
flask-swagger-ui==4.11.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
PyJWT==2.8.0
python-dotenv==1.0.0
Werkzeug==3.1.3
